SMC sweep: the physical page's linear alias is 32 bits, not 16
invalidate_for_store read phys_alias (a linear page number + 1) into a
uint16_t. V86 and UMB aliases sit below 1 MB and fit; NT's kernel code
is at 80000000h and up (win32k at A0000000h), so a store to one of its
code pages swept the wrong linear address and the stale translation
kept running. NT 4's GUI setup blue-screened at random (STOP 50h in
win32k, STOP 1Eh in Fastfat) under the JIT and never on the
interpreter; with the alias whole it reaches the setup wizard on every
timing tried.

+++ dbt/dbt_cache.c
@@ -185,7 +185,7 @@ static void invalidate_for_store(x86_dbt *dbt, uint32_t phys) {
     if (dbt->smc_win[phys] != now) { dbt->smc_win[phys] = now; dbt->smc_heat[phys] = 0; }
     if (dbt->smc_heat[phys] < 255) dbt->smc_heat[phys]++;
     sweep_blocks_at(dbt, phys);
-    uint16_t alias = dbt->phys_alias[phys >> 12];
+    uint32_t alias = dbt->phys_alias[phys >> 12];   /* a linear page + 1: past 16 bits for NT's kernel at 80000000h */
     if (alias) sweep_blocks_at(dbt, ((uint32_t)(alias - 1) << 12) | (phys & 0xFFF));
     dbt->smc_invalidations++;
     dbt->cpu->code_bitmap[phys] &= (uint8_t)~X86_BM_CODE;   /* a device bit stays */
